data-pipeline/requirements.txt
pymongo==4.6.1
sentence-transformers==2.2.2
qdrant-client==1.7.0
torch>=2.0.0
numpy>=1.24.0
tqdm>=4.65.0
python-dotenv==1.0.0
